+++ src/R/pcgrr/R/cpsr.R
@@ -4,16 +4,16 @@
 #' @param project_directory name of project directory
 #' @param query_vcf2tsv name of gzipped TSV file (vcf2tsv) with annotated query SNVs/InDels
 #' @param pcgr_data List of data frames with PCGR data annotations
-#' @param pcgr_config Object with CPSR configuration parameters
+#' @param cpsr_config Object with CPSR configuration parameters
 #' @param sample_name sample identifier
 #' @param pcgr_version PCGR software version
 #' @param genome_assembly human genome assembly version
 #'
 
-generate_predisposition_report <- function(project_directory, query_vcf2tsv, pcgr_data, pcgr_config = NULL, sample_name = "SampleX",
+generate_predisposition_report <- function(project_directory, query_vcf2tsv, pcgr_data, cpsr_config = NULL, sample_name = "SampleX",
                             pcgr_version = "0.1.0", genome_assembly = "grch37"){
 
-  pcg_report <- pcgrr::init_pcg_report(pcgr_config, sample_name, pcgr_version, genome_assembly, class = NULL, pcgr_data = pcgr_data, type = "predisposition")
+  cps_report <- pcgrr::init_pcg_report(cpsr_config, sample_name, pcgr_version, genome_assembly, class = NULL, pcgr_data = pcgr_data, type = "predisposition")
 
   genome_seq <- BSgenome.Hsapiens.UCSC.hg38
   assembly <- "hg38"
@@ -28,10 +28,10 @@ generate_predisposition_report <- function(project_directory, query_vcf2tsv, pcg
 
   ## define tags/variables to display in data tables (tier 1,2,3A,3B)
   predispose_tier1_2_display <- c("SYMBOL", "CONSEQUENCE", "PROTEIN_CHANGE", "CLINVAR_PHENOTYPE", "GENOTYPE", "GENE_NAME", "PROTEIN_DOMAIN",
-                          "HGVSp", "HGVSc", "CDS_CHANGE", "MUTATION_HOTSPOT", "RMSK_HIT","PROTEIN_FEATURE", "PREDICTED_EFFECT", "LOSS_OF_FUNCTION", "DBSNP", "CLINVAR",
+                          "HGVSp", "HGVSc", "CDS_CHANGE", "REFSEQ_MRNA","MUTATION_HOTSPOT", "RMSK_HIT","PROTEIN_FEATURE", "PREDICTED_EFFECT", "LOSS_OF_FUNCTION", "DBSNP", "CLINVAR",
                           "CLINVAR_CLINICAL_SIGNIFICANCE", "CLINVAR_VARIANT_ORIGIN", "GWAS_CITATION","ONCOGENE", "ONCOSCORE", "TUMOR_SUPPRESSOR",
-                          "GLOBAL_AF_GNOMAD", pcg_report[["pcgr_config"]][["popgen"]][["vcftag_gnomad"]], "GLOBAL_AF_1KG",
-                          pcg_report[["pcgr_config"]][["popgen"]][["vcftag_tgp"]], "GENOMIC_CHANGE", "GENOME_VERSION")
+                          "GLOBAL_AF_GNOMAD", cps_report[["cpsr_config"]][["popgen"]][["vcftag_gnomad"]], "GLOBAL_AF_1KG",
+                          cps_report[["cpsr_config"]][["popgen"]][["vcftag_tgp"]], "GENOMIC_CHANGE", "GENOME_VERSION")
 
   predispose_tier3A_display <- unique(c("SYMBOL", "CONSEQUENCE", "PROTEIN_CHANGE", "CLINVAR_PHENOTYPE", "PATHRANK", "GENOTYPE",
                                         "PATHDOC", "PATHSCORE", predispose_tier1_2_display))
@@ -42,10 +42,10 @@ generate_predisposition_report <- function(project_directory, query_vcf2tsv, pcg
   predispose_tier3B_display <- unique(c("SYMBOL","CONSEQUENCE", "PROTEIN_CHANGE","PATHRANK","PATHSCORE","GENOTYPE","PATHDOC", predispose_tier3B_display))
 
   predispose_gwas_display <- c("SYMBOL","CONSEQUENCE", "GWAS_CITATION","PROTEIN_CHANGE","GENOTYPE","LOSS_OF_FUNCTION",
-                                 "PROTEIN_CHANGE","GENE_NAME", "GWAS_PHENOTYPE","PROTEIN_DOMAIN","HGVSp", "HGVSc", "CDS_CHANGE", "PROTEIN_FEATURE", "PREDICTED_EFFECT",
+                                 "PROTEIN_CHANGE","GENE_NAME", "GWAS_PHENOTYPE","PROTEIN_DOMAIN","HGVSp", "HGVSc", "CDS_CHANGE", "REFSEQ_MRNA","PROTEIN_FEATURE", "PREDICTED_EFFECT",
                                  "DBSNP","CLINVAR","CLINVAR_PHENOTYPE","CLINVAR_CLINICAL_SIGNIFICANCE","GLOBAL_AF_GNOMAD",
-                                  pcg_report[["pcgr_config"]][["popgen"]][["vcftag_gnomad"]],
-                                 "GLOBAL_AF_1KG",pcg_report[["pcgr_config"]][["popgen"]][["vcftag_tgp"]],"GENOMIC_CHANGE","GENOME_VERSION")
+                                  cps_report[["cpsr_config"]][["popgen"]][["vcftag_gnomad"]],
+                                 "GLOBAL_AF_1KG",cps_report[["cpsr_config"]][["popgen"]][["vcftag_tgp"]],"GENOMIC_CHANGE","GENOME_VERSION")
 
 
 
@@ -53,19 +53,19 @@ generate_predisposition_report <- function(project_directory, query_vcf2tsv, pcg
     if (!file.exists(query_vcf2tsv) | file.size(query_vcf2tsv) == 0){
       rlogging::warning(paste0("File ", query_vcf2tsv, " does not exist or has zero size"))
     }else{
-      if (!is.null(pcgr_config) & query_vcf2tsv != "None.gz"){
+      if (!is.null(cpsr_config) & query_vcf2tsv != "None.gz"){
 
         ##read calls
-        calls <- pcgrr::get_calls(query_vcf2tsv, pcgr_data, pcgr_version, sample_name, pcgr_config, genome_seq, assembly)
+        calls <- pcgrr::get_calls(query_vcf2tsv, pcgr_data, pcgr_version, sample_name, cpsr_config, genome_seq, assembly)
         rlogging::message("Filtering against cancer predisposition genes:")
-        sample_calls <- dplyr::semi_join(calls, pcg_report[["snv_indel"]][["predisposition_genes"]], by = c("SYMBOL" = "symbol"))
+        sample_calls <- dplyr::semi_join(calls, cps_report[["snv_indel"]][["predisposition_genes"]], by = c("SYMBOL" = "symbol"))
         sample_calls <- dplyr::rename(sample_calls, CLINVAR_CLINICAL_SIGNIFICANCE = CLINVAR_CLNSIG)
         rlogging::message(nrow(sample_calls), " variants remaining")
         sample_calls_coding <- dplyr::filter(sample_calls, CODING_STATUS == "coding")
         gene_hits_coding <- paste(unique(sample_calls_coding$SYMBOL), collapse = ", ")
         rlogging::message("Found coding variants in the following cancer predisposition genes: ", gene_hits_coding)
 
-        pcg_report$snv_indel$eval <- TRUE
+        cps_report$snv_indel$eval <- TRUE
 
 
         rlogging::message("Limiting variants to hereditary cancer-predisposing syndromes/cancer conditions")
@@ -106,110 +106,110 @@ generate_predisposition_report <- function(project_directory, query_vcf2tsv, pcg
           vus_clnsig_regex <- "^(uncertain_significance|risk_factor|uncertain_significance,_other)$"
           all_hits <- data.frame()
 
-          ## Assign TIER 1 variants to pcg_report object
+          ## Assign TIER 1 variants to cps_report object
           for (ph in c("cancer_phenotype", "noncancer_phenotype")){
-            pcg_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]] <-  dplyr::filter(sample_calls_all[[ph]], CLINVAR_CONFLICTED == 0 & stringr::str_detect(CLINVAR_VARIANT_ORIGIN,"germline") & !is.na(CLINVAR_CLINICAL_SIGNIFICANCE) & stringr::str_detect(CLINVAR_CLINICAL_SIGNIFICANCE, pathogenic_clnsig_regex))
-            if(nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]]) > 0){
-              pcg_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]] <-
-                pcg_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]] %>%
+            cps_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]] <-  dplyr::filter(sample_calls_all[[ph]], CLINVAR_CONFLICTED == 0 & stringr::str_detect(CLINVAR_VARIANT_ORIGIN,"germline") & !is.na(CLINVAR_CLINICAL_SIGNIFICANCE) & stringr::str_detect(CLINVAR_CLINICAL_SIGNIFICANCE, pathogenic_clnsig_regex))
+            if(nrow(cps_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]]) > 0){
+              cps_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]] <-
+                cps_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]] %>%
                 dplyr::mutate(PATHRANK = NA, PATHSCORE = NA, PATHDOC = NA)
-              all_hits <- dplyr::bind_rows(all_hits, dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]],GENOMIC_CHANGE))
+              all_hits <- dplyr::bind_rows(all_hits, dplyr::select(cps_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]],GENOMIC_CHANGE))
             }
-            rlogging::message(paste0("TIER 1: Pathogenic variants - ",ph,": n = ",nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]])))
+            rlogging::message(paste0("TIER 1: Pathogenic variants - ",ph,": n = ",nrow(cps_report[["snv_indel"]][["variant_display"]][["tier1"]][[ph]])))
 
           }
 
-          ## Assign TIER 2 variants (likely pathogenic) to pcg_report object
+          ## Assign TIER 2 variants (likely pathogenic) to cps_report object
           for (ph in c("cancer_phenotype", "noncancer_phenotype")){
-            pcg_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]] <-  dplyr::filter(sample_calls_all[[ph]], CLINVAR_CONFLICTED == 0 & stringr::str_detect(CLINVAR_VARIANT_ORIGIN,"germline") & !is.na(CLINVAR_CLINICAL_SIGNIFICANCE) & stringr::str_detect(CLINVAR_CLINICAL_SIGNIFICANCE, likely_pathogenic_clnsig_regex))
-            if(nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]]) > 0){
-              pcg_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]] <-
-                pcg_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]] %>%
+            cps_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]] <-  dplyr::filter(sample_calls_all[[ph]], CLINVAR_CONFLICTED == 0 & stringr::str_detect(CLINVAR_VARIANT_ORIGIN,"germline") & !is.na(CLINVAR_CLINICAL_SIGNIFICANCE) & stringr::str_detect(CLINVAR_CLINICAL_SIGNIFICANCE, likely_pathogenic_clnsig_regex))
+            if(nrow(cps_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]]) > 0){
+              cps_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]] <-
+                cps_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]] %>%
                 dplyr::mutate(PATHRANK = NA, PATHSCORE = NA, PATHDOC = NA)
-              all_hits <- dplyr::bind_rows(all_hits, dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]],GENOMIC_CHANGE))
+              all_hits <- dplyr::bind_rows(all_hits, dplyr::select(cps_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]],GENOMIC_CHANGE))
             }
-            rlogging::message(paste0("TIER 2: Likely pathogenic variants - ",ph,": n = ",nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]])))
+            rlogging::message(paste0("TIER 2: Likely pathogenic variants - ",ph,": n = ",nrow(cps_report[["snv_indel"]][["variant_display"]][["tier2"]][[ph]])))
           }
 
-          ## Assign TIER 3A (VUS in ClinVar) variants to pcg_report object
+          ## Assign TIER 3A (VUS in ClinVar) variants to cps_report object
           for (ph in c("cancer_phenotype", "noncancer_phenotype")){
-            pcg_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]] <-  dplyr::filter(sample_calls_all[[ph]], CLINVAR_CONFLICTED == 0 & stringr::str_detect(CLINVAR_VARIANT_ORIGIN, "germline") & !is.na(CLINVAR_CLINICAL_SIGNIFICANCE) & stringr::str_detect(CLINVAR_CLINICAL_SIGNIFICANCE, vus_clnsig_regex))
-            if(nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]]) > 0){
-              pcg_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]] <-
-                pcgrr::assign_pathogenicity_score(pcg_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]], pcgr_config, pcgr_data)
-              pcg_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]] <-
-                pcg_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]] %>%
+            cps_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]] <-  dplyr::filter(sample_calls_all[[ph]], CLINVAR_CONFLICTED == 0 & stringr::str_detect(CLINVAR_VARIANT_ORIGIN, "germline") & !is.na(CLINVAR_CLINICAL_SIGNIFICANCE) & stringr::str_detect(CLINVAR_CLINICAL_SIGNIFICANCE, vus_clnsig_regex))
+            if(nrow(cps_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]]) > 0){
+              cps_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]] <-
+                pcgrr::assign_pathogenicity_score(cps_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]], cpsr_config, pcgr_data)
+              cps_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]] <-
+                cps_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]] %>%
                 dplyr::arrange(desc(PATHSCORE), LOSS_OF_FUNCTION, CODING_STATUS, desc(ONCOSCORE))
-              all_hits <- dplyr::bind_rows(all_hits, dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]],GENOMIC_CHANGE))
+              all_hits <- dplyr::bind_rows(all_hits, dplyr::select(cps_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]],GENOMIC_CHANGE))
             }
-            rlogging::message(paste0("TIER 3: Variants of uncertain significance - ",ph,": n = ",nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]])))
+            rlogging::message(paste0("TIER 3: Variants of uncertain significance - ",ph,": n = ",nrow(cps_report[["snv_indel"]][["variant_display"]][["tier3A"]][[ph]])))
           }
 
-          ## Assign TIER 3B (Non-classified variants) to pcg_report object
-          pcg_report[["snv_indel"]][["variant_display"]][["tier3B"]] <-  dplyr::filter(sample_calls_all[["noncancer_phenotype"]], is.na(CLINVAR_CLINICAL_SIGNIFICANCE) & CODING_STATUS == "coding" & (is.na(GLOBAL_AF_GNOMAD) | GLOBAL_AF_GNOMAD < pcgr_config$maf_limits$maf_gnomad) & (is.na(GLOBAL_AF_1KG) | GLOBAL_AF_1KG < pcgr_config$maf_limits$maf_tgp))
-          if(nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier3B"]]) > 0){
-            pcg_report[["snv_indel"]][["variant_display"]][["tier3B"]] <-
-              pcgrr::assign_pathogenicity_score(pcg_report[["snv_indel"]][["variant_display"]][["tier3B"]], pcgr_config, pcgr_data)
-            pcg_report[["snv_indel"]][["variant_display"]][["tier3B"]] <-
-              pcg_report[["snv_indel"]][["variant_display"]][["tier3B"]] %>%
+          ## Assign TIER 3B (Non-classified variants) to cps_report object
+          cps_report[["snv_indel"]][["variant_display"]][["tier3B"]] <-  dplyr::filter(sample_calls_all[["noncancer_phenotype"]], is.na(CLINVAR_CLINICAL_SIGNIFICANCE) & CODING_STATUS == "coding" & (is.na(GLOBAL_AF_GNOMAD) | GLOBAL_AF_GNOMAD < cpsr_config$maf_limits$maf_gnomad) & (is.na(GLOBAL_AF_1KG) | GLOBAL_AF_1KG < cpsr_config$maf_limits$maf_tgp))
+          if(nrow(cps_report[["snv_indel"]][["variant_display"]][["tier3B"]]) > 0){
+            cps_report[["snv_indel"]][["variant_display"]][["tier3B"]] <-
+              pcgrr::assign_pathogenicity_score(cps_report[["snv_indel"]][["variant_display"]][["tier3B"]], cpsr_config, pcgr_data)
+            cps_report[["snv_indel"]][["variant_display"]][["tier3B"]] <-
+              cps_report[["snv_indel"]][["variant_display"]][["tier3B"]] %>%
               dplyr::arrange(desc(PATHSCORE), LOSS_OF_FUNCTION, CODING_STATUS, desc(ONCOSCORE))
-            all_hits <- dplyr::bind_rows(all_hits, dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][["tier3B"]],GENOMIC_CHANGE))
+            all_hits <- dplyr::bind_rows(all_hits, dplyr::select(cps_report[["snv_indel"]][["variant_display"]][["tier3B"]],GENOMIC_CHANGE))
           }
-          rlogging::message(paste0("TIER 3: Other unclassified variants - ",ph,": n = ",nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier3B"]])))
+          rlogging::message(paste0("TIER 3: Other unclassified variants - ",ph,": n = ",nrow(cps_report[["snv_indel"]][["variant_display"]][["tier3B"]])))
 
-          if(pcgr_config[['gwas']][['gwas_hits']] == TRUE){
+          if(cpsr_config[['gwas']][['gwas_hits']] == TRUE){
             rlogging::message("Assignment of other variants to hits from genome-wide association studies")
           }
-          ## Assign GWAS hits to pcg_report object
-          pcg_report[["snv_indel"]][["variant_display"]][["gwas"]] <-  dplyr::filter(calls, !is.na(GWAS_HIT) & !is.na(GWAS_CITATION))
-          pcg_report[["snv_indel"]][["variant_display"]][["gwas"]] <- dplyr::anti_join(pcg_report[["snv_indel"]][["variant_display"]][["gwas"]], all_hits, by=c("GENOMIC_CHANGE"))
-          if(nrow(pcg_report[["snv_indel"]][["variant_display"]][["gwas"]]) > 0){
-            pcg_report[["snv_indel"]][["variant_display"]][["gwas"]] <-
-              pcg_report[["snv_indel"]][["variant_display"]][["gwas"]] %>%
+          ## Assign GWAS hits to cps_report object
+          cps_report[["snv_indel"]][["variant_display"]][["gwas"]] <-  dplyr::filter(calls, !is.na(GWAS_HIT) & !is.na(GWAS_CITATION))
+          cps_report[["snv_indel"]][["variant_display"]][["gwas"]] <- dplyr::anti_join(cps_report[["snv_indel"]][["variant_display"]][["gwas"]], all_hits, by=c("GENOMIC_CHANGE"))
+          if(nrow(cps_report[["snv_indel"]][["variant_display"]][["gwas"]]) > 0){
+            cps_report[["snv_indel"]][["variant_display"]][["gwas"]] <-
+              cps_report[["snv_indel"]][["variant_display"]][["gwas"]] %>%
               dplyr::mutate(PATHRANK = NA, PATHSCORE = NA, PATHDOC = NA, CLINVAR_PHENOTYPE = NA) %>%
               dplyr::rename(CLINVAR_CLINICAL_SIGNIFICANCE = CLINVAR_CLNSIG) %>%
               dplyr::arrange(LOSS_OF_FUNCTION, CODING_STATUS, desc(ONCOSCORE))
           }
-          if(pcgr_config[['gwas']][['gwas_hits']] == TRUE){
-            rlogging::message(paste0("GWAS hits - cancer phenotypes: n = ",nrow(pcg_report[["snv_indel"]][["variant_display"]][["gwas"]])))
+          if(cpsr_config[['gwas']][['gwas_hits']] == TRUE){
+            rlogging::message(paste0("GWAS hits - cancer phenotypes: n = ",nrow(cps_report[["snv_indel"]][["variant_display"]][["gwas"]])))
           }
 
-          pcg_report <- pcgrr::generate_tier_tsv_cpsr(pcg_report, sample_name = sample_name)
-          pcg_report <- pcgrr::summary_findings_cpsr(pcg_report)
-          population_tags <- unique(c("GLOBAL_AF_1KG", pcg_report[["pcgr_config"]][["popgen"]][["vcftag_tgp"]],
-                                      "GLOBAL_AF_GNOMAD", pcg_report[["pcgr_config"]][["popgen"]][["vcftag_gnomad"]]))
+          cps_report <- pcgrr::generate_tier_tsv_cpsr(cps_report, sample_name = sample_name)
+          cps_report <- pcgrr::summary_findings_cpsr(cps_report)
+          population_tags <- unique(c("GLOBAL_AF_1KG", cps_report[["cpsr_config"]][["popgen"]][["vcftag_tgp"]],
+                                      "GLOBAL_AF_GNOMAD", cps_report[["cpsr_config"]][["popgen"]][["vcftag_gnomad"]]))
           for (class in c("tier1", "tier2", "tier3A", "tier3B","gwas")){
             if (class != "tier3B" & class != "gwas"){
               for (c in c("noncancer_phenotype", "cancer_phenotype")){
-                if (nrow(pcg_report[["snv_indel"]][["variant_display"]][[class]][[c]]) > 0){
-                  pcg_report[["snv_indel"]][["variant_display"]][[class]][[c]] <- pcg_report[["snv_indel"]][["variant_display"]][[class]][[c]] %>%
+                if (nrow(cps_report[["snv_indel"]][["variant_display"]][[class]][[c]]) > 0){
+                  cps_report[["snv_indel"]][["variant_display"]][[class]][[c]] <- cps_report[["snv_indel"]][["variant_display"]][[class]][[c]] %>%
                     dplyr::mutate(CLINVAR = paste0("<a href=\"http://www.ncbi.nlm.nih.gov/clinvar/variation/", CLINVAR_MSID, "\" target=\"_blank\">", CLINVAR_MSID, "</a>")) %>%
                     dplyr::select(-CLINVAR_MSID)
                   for (tag in population_tags){
-                    if (nrow(pcg_report[["snv_indel"]][["variant_display"]][[class]][[c]][is.na(pcg_report[["snv_indel"]][["variant_display"]][[class]][[c]][, tag]), ]) > 0){
-                      pcg_report[["snv_indel"]][["variant_display"]][[class]][[c]][is.na(pcg_report[["snv_indel"]][["variant_display"]][[class]][[c]][, tag]), ][, tag] <- 0.00
+                    if (nrow(cps_report[["snv_indel"]][["variant_display"]][[class]][[c]][is.na(cps_report[["snv_indel"]][["variant_display"]][[class]][[c]][, tag]), ]) > 0){
+                      cps_report[["snv_indel"]][["variant_display"]][[class]][[c]][is.na(cps_report[["snv_indel"]][["variant_display"]][[class]][[c]][, tag]), ][, tag] <- 0.00
                     }
                   }
                   if (class == "tier1" | class == "tier2"){
-                    pcg_report[["snv_indel"]][["variant_display"]][[class]][[c]] <- dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][[class]][[c]], dplyr::one_of(predispose_tier1_2_display))
+                    cps_report[["snv_indel"]][["variant_display"]][[class]][[c]] <- dplyr::select(cps_report[["snv_indel"]][["variant_display"]][[class]][[c]], dplyr::one_of(predispose_tier1_2_display))
                   }
                   else{
-                    pcg_report[["snv_indel"]][["variant_display"]][[class]][[c]] <- dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][[class]][[c]], dplyr::one_of(predispose_tier3A_display))
+                    cps_report[["snv_indel"]][["variant_display"]][[class]][[c]] <- dplyr::select(cps_report[["snv_indel"]][["variant_display"]][[class]][[c]], dplyr::one_of(predispose_tier3A_display))
                   }
                 }
               }
             }else{
-              if (nrow(pcg_report[["snv_indel"]][["variant_display"]][[class]]) > 0){
+              if (nrow(cps_report[["snv_indel"]][["variant_display"]][[class]]) > 0){
                 for (tag in population_tags){
-                  if (nrow(pcg_report[["snv_indel"]][["variant_display"]][[class]][is.na(pcg_report[["snv_indel"]][["variant_display"]][[class]][, tag]), ]) > 0){
-                    pcg_report[["snv_indel"]][["variant_display"]][[class]][is.na(pcg_report[["snv_indel"]][["variant_display"]][[class]][, tag]), ][, tag] <- 0.00
+                  if (nrow(cps_report[["snv_indel"]][["variant_display"]][[class]][is.na(cps_report[["snv_indel"]][["variant_display"]][[class]][, tag]), ]) > 0){
+                    cps_report[["snv_indel"]][["variant_display"]][[class]][is.na(cps_report[["snv_indel"]][["variant_display"]][[class]][, tag]), ][, tag] <- 0.00
                   }
                 }
                 if(class == "tier3B"){
-                  pcg_report[["snv_indel"]][["variant_display"]][[class]] <- dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][[class]], dplyr::one_of(predispose_tier3B_display))
+                  cps_report[["snv_indel"]][["variant_display"]][[class]] <- dplyr::select(cps_report[["snv_indel"]][["variant_display"]][[class]], dplyr::one_of(predispose_tier3B_display))
                 }
                 else{
-                  pcg_report[["snv_indel"]][["variant_display"]][[class]] <- dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][[class]], dplyr::one_of(predispose_gwas_display))
+                  cps_report[["snv_indel"]][["variant_display"]][[class]] <- dplyr::select(cps_report[["snv_indel"]][["variant_display"]][[class]], dplyr::one_of(predispose_gwas_display))
 
                 }
               }
@@ -220,28 +220,28 @@ generate_predisposition_report <- function(project_directory, query_vcf2tsv, pcg
     }
 
     fname_key <- "tsv"
-    if (!is.null(pcg_report[["snv_indel"]][["variant_set"]][[fname_key]])){
-      if (nrow(pcg_report[["snv_indel"]][["variant_set"]][[fname_key]]) > 0){
-        write.table(pcg_report[["snv_indel"]][["variant_set"]][[fname_key]], file = fnames[[fname_key]], sep = "\t", col.names = T, row.names = F, quote = F)
+    if (!is.null(cps_report[["snv_indel"]][["variant_set"]][[fname_key]])){
+      if (nrow(cps_report[["snv_indel"]][["variant_set"]][[fname_key]]) > 0){
+        write.table(cps_report[["snv_indel"]][["variant_set"]][[fname_key]], file = fnames[[fname_key]], sep = "\t", col.names = T, row.names = F, quote = F)
       }
     }
 
-    pcgr_json <- jsonlite::toJSON(pcg_report, pretty = T, na = "string", null = "null")
+    pcgr_json <- jsonlite::toJSON(cps_report, pretty = T, na = "string", null = "null")
     write(pcgr_json, fnames[["json"]])
     gzip_command <- paste0("gzip -f ", fnames[["json"]])
     system(gzip_command, intern = F)
-    rmarkdown::render(system.file("templates", "report_predisposition.Rmd", package = "pcgrr"), output_format = rmarkdown::html_document(theme = pcg_report[["pcgr_config"]][["visual"]][["report_theme"]], toc = T, toc_depth = 3, toc_float = T, number_sections = F, includes = rmarkdown::includes(after_body = "disclaimer_predisposition.md")), output_file = paste0(sample_name, ".cpsr.", genome_assembly, ".html"), output_dir = project_directory, clean = T, intermediates_dir = project_directory, quiet = T)
+    rmarkdown::render(system.file("templates", "report_predisposition.Rmd", package = "pcgrr"), output_format = rmarkdown::html_document(theme = cps_report[["cpsr_config"]][["visual"]][["report_theme"]], toc = T, toc_depth = 3, toc_float = T, number_sections = F, includes = rmarkdown::includes(after_body = "disclaimer_predisposition.md")), output_file = paste0(sample_name, ".cpsr.", genome_assembly, ".html"), output_dir = project_directory, clean = T, intermediates_dir = project_directory, quiet = T)
   }
 }
 
 #' Function that generates summary findings for CPSR
 #'
-#' @param pcg_report report object
+#' @param cps_report report object
 #'
-summary_findings_cpsr <- function(pcg_report){
+summary_findings_cpsr <- function(cps_report){
 
-  if(nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier1"]][["cancer_phenotype"]]) > 0){
-    pcg_report[["summary"]][["tier1"]] <- as.data.frame(dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][["tier1"]][["cancer_phenotype"]], CLINVAR_MSID, CLINVAR_PHENOTYPE, SYMBOL, HGVSp_short, CONSEQUENCE, ONCOSCORE) %>%
+  if(nrow(cps_report[["snv_indel"]][["variant_display"]][["tier1"]][["cancer_phenotype"]]) > 0){
+    cps_report[["summary"]][["tier1"]] <- as.data.frame(dplyr::select(cps_report[["snv_indel"]][["variant_display"]][["tier1"]][["cancer_phenotype"]], CLINVAR_MSID, CLINVAR_PHENOTYPE, SYMBOL, HGVSp_short, CONSEQUENCE, ONCOSCORE) %>%
       dplyr::mutate(MUTATION = paste0("<a href=\"http://www.ncbi.nlm.nih.gov/clinvar/variation/",CLINVAR_MSID, "\" target=\"_blank\">", SYMBOL,":",CONSEQUENCE,":",HGVSp_short,"</a>")) %>%
       dplyr::mutate(MUTATION = stringr::str_replace(MUTATION,":NA$","")) %>%
       dplyr::select(-c(HGVSp_short,SYMBOL,CONSEQUENCE, CLINVAR_MSID)) %>%
@@ -253,8 +253,8 @@ summary_findings_cpsr <- function(pcg_report){
       head(2)
   }
 
-  if(nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier2"]][["cancer_phenotype"]]) > 0){
-    pcg_report[["summary"]][["tier2"]] <- as.data.frame(dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][["tier2"]][["cancer_phenotype"]], CLINVAR_MSID, CLINVAR_PHENOTYPE, SYMBOL, HGVSp_short, CONSEQUENCE, ONCOSCORE) %>%
+  if(nrow(cps_report[["snv_indel"]][["variant_display"]][["tier2"]][["cancer_phenotype"]]) > 0){
+    cps_report[["summary"]][["tier2"]] <- as.data.frame(dplyr::select(cps_report[["snv_indel"]][["variant_display"]][["tier2"]][["cancer_phenotype"]], CLINVAR_MSID, CLINVAR_PHENOTYPE, SYMBOL, HGVSp_short, CONSEQUENCE, ONCOSCORE) %>%
       dplyr::mutate(MUTATION = paste0("<a href=\"http://www.ncbi.nlm.nih.gov/clinvar/variation/",CLINVAR_MSID, "\" target=\"_blank\">", SYMBOL,":",CONSEQUENCE,":",HGVSp_short,"</a>")) %>%
       dplyr::mutate(MUTATION = stringr::str_replace(MUTATION,":NA$","")) %>%
       dplyr::select(-c(HGVSp_short,SYMBOL,CONSEQUENCE,CLINVAR_MSID)) %>%
@@ -266,36 +266,36 @@ summary_findings_cpsr <- function(pcg_report){
       head(2)
   }
 
-  if(nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier3A"]][["cancer_phenotype"]]) > 0){
-    pcg_report[["summary"]][["tier3A"]] <- dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][["tier3A"]][["cancer_phenotype"]], CLINVAR_MSID, CLINVAR_PHENOTYPE, SYMBOL, PATHRANK, HGVSp_short, CONSEQUENCE, ONCOSCORE) %>%
+  if(nrow(cps_report[["snv_indel"]][["variant_display"]][["tier3A"]][["cancer_phenotype"]]) > 0){
+    cps_report[["summary"]][["tier3A"]] <- dplyr::select(cps_report[["snv_indel"]][["variant_display"]][["tier3A"]][["cancer_phenotype"]], CLINVAR_MSID, CLINVAR_PHENOTYPE, SYMBOL, PATHRANK, HGVSp_short, CONSEQUENCE, ONCOSCORE) %>%
       dplyr::mutate(MUTATION = paste0("<a href=\"http://www.ncbi.nlm.nih.gov/clinvar/variation/",CLINVAR_MSID, "\" target=\"_blank\">", SYMBOL,":",CONSEQUENCE,":",HGVSp_short,"</a>")) %>%
       dplyr::mutate(MUTATION = stringr::str_replace(MUTATION,":NA$","")) %>%
       dplyr::filter(PATHRANK == 'HIGH')
 
-      if(nrow(pcg_report[["summary"]][["tier3A"]]) > 0){
-        pcg_report[["summary"]][["tier3A"]] <- pcg_report[["summary"]][["tier3A"]] %>%
+      if(nrow(cps_report[["summary"]][["tier3A"]]) > 0){
+        cps_report[["summary"]][["tier3A"]] <- cps_report[["summary"]][["tier3A"]] %>%
           dplyr::select(-c(HGVSp_short,SYMBOL,CONSEQUENCE, CLINVAR_MSID)) %>%
           tidyr::separate_rows(CLINVAR_PHENOTYPE,sep="; ") %>%
           dplyr::filter(CLINVAR_PHENOTYPE != 'not provided' & CLINVAR_PHENOTYPE != 'not specified')
       }
 
-      if(nrow(pcg_report[["summary"]][["tier3A"]]) > 0){
-        pcg_report[["summary"]][["tier3A"]] <- pcg_report[["summary"]][["tier3A"]] %>%
+      if(nrow(cps_report[["summary"]][["tier3A"]]) > 0){
+        cps_report[["summary"]][["tier3A"]] <- cps_report[["summary"]][["tier3A"]] %>%
           dplyr::group_by(CLINVAR_PHENOTYPE) %>%
           dplyr::summarise(VARIANTS = paste(unique(MUTATION), collapse=", "), n = n()) %>%
           dplyr::arrange(desc(n))
       }
   }
 
-  if(nrow(pcg_report[["snv_indel"]][["variant_display"]][["tier3B"]]) > 0){
-    pcg_report[["summary"]][["tier3B"]] <- dplyr::select(pcg_report[["snv_indel"]][["variant_display"]][["tier3B"]],  SYMBOL, PATHRANK, HGVSp_short, CONSEQUENCE, ONCOSCORE) %>%
+  if(nrow(cps_report[["snv_indel"]][["variant_display"]][["tier3B"]]) > 0){
+    cps_report[["summary"]][["tier3B"]] <- dplyr::select(cps_report[["snv_indel"]][["variant_display"]][["tier3B"]],  SYMBOL, PATHRANK, HGVSp_short, CONSEQUENCE, ONCOSCORE) %>%
         dplyr::mutate(MUTATION =paste0(SYMBOL,":",CONSEQUENCE,":",HGVSp_short)) %>%
         dplyr::mutate(MUTATION = stringr::str_replace(MUTATION,":NA$","")) %>%
         dplyr::select(-c(HGVSp_short,SYMBOL,CONSEQUENCE)) %>%
         dplyr::filter(PATHRANK == 'HIGH')
   }
 
-  return(pcg_report)
+  return(cps_report)
 }
 
 
@@ -328,12 +328,12 @@ get_insilico_prediction_statistics <- function(sample_calls){
 #' Function that assigns variant pathogenicity scores based on ACMG guidelines
 #'
 #' @param sample_calls sample calls with dbnsfp annotations
-#' @param pcgr_config pcgr configuration object
+#' @param cpsr_config pcgr configuration object
 #' @param pcgr_data pcgr data object
 #'
 #' @return sample_calls
 #'
-assign_pathogenicity_score <- function(sample_calls, pcgr_config, pcgr_data){
+assign_pathogenicity_score <- function(sample_calls, cpsr_config, pcgr_data){
 
   path_columns <- c("PVS1","PSC1","PS1","PM1", "PM2", "BA1","PP3","PM4","PPC1", "BP4", "PP2","BMC1","BSC1","N_INSILICO_CALLED",
                     "N_INSILICO_DAMAGING", "N_INSILICO_TOLERATED", "N_INSILICO_SPLICING_NEUTRAL",
@@ -380,8 +380,8 @@ assign_pathogenicity_score <- function(sample_calls, pcgr_config, pcgr_data){
   #    and at most one predicted as damaging (BP4)
   #    - 0 predictions of splice site affected
 
-  dbnsfp_min_majority <- pcgr_config[['dbnsfp']][['min_majority']]
-  dbnsfp_max_minority <- pcgr_config[['dbnsfp']][['max_minority']]
+  dbnsfp_min_majority <- cpsr_config[['dbnsfp']][['min_majority']]
+  dbnsfp_max_minority <- cpsr_config[['dbnsfp']][['max_minority']]
   dbnsfp_min_called <- dbnsfp_min_majority
 
   sample_calls <- pcgrr::get_insilico_prediction_statistics(sample_calls)
@@ -445,21 +445,20 @@ assign_pathogenicity_score <- function(sample_calls, pcgr_config, pcgr_data){
       sample_calls[!is.na(sample_calls$LOSS_OF_FUNCTION) & sample_calls$LOSS_OF_FUNCTION == TRUE & !is.na(sample_calls$LOF_KNOWN_MOI) & sample_calls$LOF_KNOWN_MOI == "LoF" & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"AD|AD/AR"),]$PVS1 <- TRUE
     }
 
-    if(nrow(sample_calls[!is.na(sample_calls$LOSS_OF_FUNCTION) & sample_calls$LOSS_OF_FUNCTION == TRUE & !is.na(sample_calls$LOF_KNOWN_MOI) & sample_calls$LOF_KNOWN_MOI == "LoF" & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"^AR$"),]) > 0){
-      sample_calls[!is.na(sample_calls$LOSS_OF_FUNCTION) & sample_calls$LOSS_OF_FUNCTION == TRUE & !is.na(sample_calls$LOF_KNOWN_MOI) & sample_calls$LOF_KNOWN_MOI == "LoF" & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"^AR$"),]$PSC1 <- TRUE
-      ## skip genes that are associated with gain-of-function
+    if(nrow(sample_calls[!is.na(sample_calls$LOSS_OF_FUNCTION) & sample_calls$LOSS_OF_FUNCTION == TRUE & !is.na(sample_calls$LOF_KNOWN_MOI) & sample_calls$LOF_KNOWN_MOI == "LoF" & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & !stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"AD|Mosaic"),]) > 0){
+      sample_calls[!is.na(sample_calls$LOSS_OF_FUNCTION) & sample_calls$LOSS_OF_FUNCTION == TRUE & !is.na(sample_calls$LOF_KNOWN_MOI) & sample_calls$LOF_KNOWN_MOI == "LoF" & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & !stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"AD|Mosaic"),]$PSC1 <- TRUE
     }
   }
 
   ## Assign a logical ACMG level
-  # PM4 - Protein length changes due to inframe indels or nonstop variant of genes that harbor variants with a dominant mode of inheritance - PM4
-  # PCC1 - Protein length changes due to inframe indels or nonstop variant of genes that harbor variants with a recessive mode of inheritance (and unknown MOI) - PPC1
+  # PM4 - Protein length changes (in non-repetitive regions) due to inframe indels or nonstop variant of genes that harbor variants with a dominant mode of inheritance - PM4
+  # PCC1 - Protein length changes (in non-repetitive regions) due to inframe indels or nonstop variant of genes that harbor variants with a recessive mode of inheritance (and unknown MOI) - PPC1
   if(nrow(sample_calls[!is.na(sample_calls$CONSEQUENCE) & stringr::str_detect(sample_calls$CONSEQUENCE,'stop_lost|inframe_deletion|inframe_insertion'),]) > 0){
-    if(nrow(sample_calls[!is.na(sample_calls$CONSEQUENCE) & stringr::str_detect(sample_calls$CONSEQUENCE,'stop_lost|inframe_deletion|inframe_insertion') & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"AD/AR|AD"),]) > 0){
-      sample_calls[!is.na(sample_calls$CONSEQUENCE) & stringr::str_detect(sample_calls$CONSEQUENCE,'stop_lost|inframe_deletion|inframe_insertion') & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"AD/AR|AD"),]$PM4 <- TRUE
+    if(nrow(sample_calls[!is.na(sample_calls$CONSEQUENCE) & is.na(RMSK_HIT) & stringr::str_detect(sample_calls$CONSEQUENCE,'stop_lost|inframe_deletion|inframe_insertion') & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"AD/AR|AD"),]) > 0){
+      sample_calls[!is.na(sample_calls$CONSEQUENCE) & is.na(RMSK_HIT) & stringr::str_detect(sample_calls$CONSEQUENCE,'stop_lost|inframe_deletion|inframe_insertion') & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"AD/AR|AD"),]$PM4 <- TRUE
     }
-    if(nrow(sample_calls[!is.na(sample_calls$CONSEQUENCE) & stringr::str_detect(sample_calls$CONSEQUENCE,'stop_lost|inframe_deletion|inframe_insertion') & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & !stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"AD|Mosaic"),]) > 0){
-      sample_calls[!is.na(sample_calls$CONSEQUENCE) & stringr::str_detect(sample_calls$CONSEQUENCE,'stop_lost|inframe_deletion|inframe_insertion') & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & !stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"AD|Mosaic"),]$PPC1 <- TRUE
+    if(nrow(sample_calls[!is.na(sample_calls$CONSEQUENCE) & is.na(RMSK_HIT) & stringr::str_detect(sample_calls$CONSEQUENCE,'stop_lost|inframe_deletion|inframe_insertion') & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & !stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"AD|Mosaic"),]) > 0){
+      sample_calls[!is.na(sample_calls$CONSEQUENCE) & is.na(RMSK_HIT) & stringr::str_detect(sample_calls$CONSEQUENCE,'stop_lost|inframe_deletion|inframe_insertion') & !is.na(sample_calls$CANCER_PREDISPOSITION_MOI) & !stringr::str_detect(sample_calls$CANCER_PREDISPOSITION_MOI,"AD|Mosaic"),]$PPC1 <- TRUE
     }
   }
 
@@ -600,7 +599,7 @@ assign_pathogenicity_score <- function(sample_calls, pcgr_config, pcgr_data){
   }
   if(nrow(sample_calls[sample_calls$PP3 == TRUE,]) > 0){
     sample_calls[sample_calls$PP3 == TRUE,]$PATHSCORE <- sample_calls[sample_calls$PP3 == TRUE,]$PATHSCORE + 1
-    sample_calls[sample_calls$PP3 == TRUE,]$PATHDOC <- paste(sample_calls[sample_calls$PP3 == TRUE,]$PATHDOC,paste0("- Multiple lines (>=",pcgr_config[['dbnsfp']][['min_majority']],") of in silico evidence of deleterious effect - PP3"),sep="<br>")
+    sample_calls[sample_calls$PP3 == TRUE,]$PATHDOC <- paste(sample_calls[sample_calls$PP3 == TRUE,]$PATHDOC,paste0("- Multiple lines of in silico evidence of deleterious effect - PP3"),sep="<br>")
   }
 
   if(nrow(sample_calls[sample_calls$BP1 == TRUE,]) > 0){
@@ -610,7 +609,7 @@ assign_pathogenicity_score <- function(sample_calls, pcgr_config, pcgr_data){
 
   if(nrow(sample_calls[sample_calls$BP4 == TRUE,]) > 0){
     sample_calls[sample_calls$BP4 == TRUE,]$PATHSCORE <- sample_calls[sample_calls$BP4 == TRUE,]$PATHSCORE - 1
-    sample_calls[sample_calls$BP4 == TRUE,]$PATHDOC <- paste(sample_calls[sample_calls$BP4 == TRUE,]$PATHDOC,paste0("- Multiple lines (>=",pcgr_config[['dbnsfp']][['min_majority']],") of in silico evidence of benign effect - BP4"),sep="<br>")
+    sample_calls[sample_calls$BP4 == TRUE,]$PATHDOC <- paste(sample_calls[sample_calls$BP4 == TRUE,]$PATHDOC,paste0("- Multiple lines of in silico evidence of benign effect - BP4"),sep="<br>")
   }
 
   if(nrow(sample_calls[sample_calls$BMC1 == TRUE,]) > 0){
@@ -653,21 +652,21 @@ assign_pathogenicity_score <- function(sample_calls, pcgr_config, pcgr_data){
 
 #' Function that generates tiered annotated variant datasets for CPSR
 #'
-#' @param pcg_report List with tiered variants
+#' @param cps_report List with tiered variants
 #'
 #' @return tsv_variants data frame with tier-annotated list of variants for tab-separated output
 #'
-generate_tier_tsv_cpsr <- function(pcg_report, sample_name = "test"){
+generate_tier_tsv_cpsr <- function(cps_report, sample_name = "test"){
 
-  predispose_tsv_tags <- c("VAR_ID","VCF_SAMPLE_ID","CODING_STATUS","SYMBOL","ENSEMBL_GENE_ID","ENSEMBL_TRANSCRIPT_ID","GENE_NAME",
+  predispose_tsv_tags <- c("VAR_ID","VCF_SAMPLE_ID","CODING_STATUS","SYMBOL","ENSEMBL_GENE_ID","ENSEMBL_TRANSCRIPT_ID","REFSEQ_MRNA","GENE_NAME",
                            "LOF_KNOWN_MOI","PATH_TRUNCATION_RATE","BENIGN_MISSENSE_RATE","ONCOGENE", "ONCOSCORE","TUMOR_SUPPRESSOR",
                            "GENOTYPE","CONSEQUENCE","PROTEIN_CHANGE","PROTEIN_DOMAIN", "HGVSp","HGVSc", "CDS_CHANGE","MUTATION_HOTSPOT","RMSK_HIT","PROTEIN_FEATURE","EFFECT_PREDICTIONS",
                            "LOSS_OF_FUNCTION", "DBSNP","CLINVAR_CLINICAL_SIGNIFICANCE", "CLINVAR_MSID","CLINVAR_VARIANT_ORIGIN",
                            "CLINVAR_CONFLICTED", "CLINVAR_PHENOTYPE","VEP_ALL_CONSEQUENCE",
                            "PATHSCORE","PATHRANK", "PATHDOC","PVS1","PSC1","PS1","PMC1","PM1","PM2","PM4","PM5","PP2","PPC1","PP3","BP3","BP4","BMC1","BSC1","BA1","BP1",
                            "N_INSILICO_CALLED","N_INSILICO_DAMAGING","N_INSILICO_TOLERATED","N_INSILICO_SPLICING_NEUTRAL","N_INSILICO_SPLICING_AFFECTED",
-                           "GLOBAL_AF_GNOMAD", pcg_report[['pcgr_config']][['popgen']][['vcftag_gnomad']],
-                           "GLOBAL_AF_1KG", pcg_report[['pcgr_config']][['popgen']][['vcftag_tgp']],"TIER","TIER_DESCRIPTION",
+                           "GLOBAL_AF_GNOMAD", cps_report[['cpsr_config']][['popgen']][['vcftag_gnomad']],
+                           "GLOBAL_AF_1KG", cps_report[['cpsr_config']][['popgen']][['vcftag_tgp']],"TIER","TIER_DESCRIPTION",
                            "GENOMIC_CHANGE", "GENOME_VERSION")
 
   rlogging::message("Generating tiered set of result variants for output in tab-separated values (TSV) file")
@@ -678,8 +677,8 @@ generate_tier_tsv_cpsr <- function(pcg_report, sample_name = "test"){
       predispose_tags <- predispose_tsv_tags
       for(ph in c('cancer_phenotype','noncancer_phenotype')){
         tierset <- data.frame()
-        if(nrow(pcg_report[['snv_indel']][['variant_display']][[tier]][[ph]]) > 0){
-          tierset <- pcg_report[['snv_indel']][['variant_display']][[tier]][[ph]]
+        if(nrow(cps_report[['snv_indel']][['variant_display']][[tier]][[ph]]) > 0){
+          tierset <- cps_report[['snv_indel']][['variant_display']][[tier]][[ph]]
           tierset$VCF_SAMPLE_ID <- sample_name
           if(tier == 'tier1'){
             tierset$TIER <- 'TIER 1'
@@ -706,9 +705,9 @@ generate_tier_tsv_cpsr <- function(pcg_report, sample_name = "test"){
             }
           }
         }
-        if(!is.null(pcg_report[['pcgr_config']][['custom_tags']])){
-          if(pcg_report[['pcgr_config']][['custom_tags']][['custom_tags']] != ""){
-            tags <- stringr::str_split(pcg_report[['pcgr_config']][['custom_tags']][['custom_tags']],pattern = ",")[[1]]
+        if(!is.null(cps_report[['cpsr_config']][['custom_tags']])){
+          if(cps_report[['cpsr_config']][['custom_tags']][['custom_tags']] != ""){
+            tags <- stringr::str_split(cps_report[['cpsr_config']][['custom_tags']][['custom_tags']],pattern = ",")[[1]]
             for(t in tags){
               t <- stringr::str_trim(t)
               if(t %in% colnames(tierset)){
@@ -723,9 +722,9 @@ generate_tier_tsv_cpsr <- function(pcg_report, sample_name = "test"){
       }
     }else{
       predispose_tags <- predispose_tsv_tags
-      if(!is.null(pcg_report[['pcgr_config']][['custom_tags']])){
-        if(pcg_report[['pcgr_config']][['custom_tags']][['custom_tags']] != ""){
-          tags <- stringr::str_split(pcg_report[['pcgr_config']][['custom_tags']][['custom_tags']],pattern = ",")[[1]]
+      if(!is.null(cps_report[['cpsr_config']][['custom_tags']])){
+        if(cps_report[['cpsr_config']][['custom_tags']][['custom_tags']] != ""){
+          tags <- stringr::str_split(cps_report[['cpsr_config']][['custom_tags']][['custom_tags']],pattern = ",")[[1]]
           for(t in tags){
             t <- stringr::str_trim(t)
             if(t %in% colnames(tierset)){
@@ -736,8 +735,8 @@ generate_tier_tsv_cpsr <- function(pcg_report, sample_name = "test"){
       }
 
       #cat(predispose_tags,'\n')
-      if(nrow(pcg_report[['snv_indel']][['variant_display']][[tier]]) > 0){
-        tierset <- pcg_report[['snv_indel']][['variant_display']][[tier]]
+      if(nrow(cps_report[['snv_indel']][['variant_display']][[tier]]) > 0){
+        tierset <- cps_report[['snv_indel']][['variant_display']][[tier]]
         tierset$VCF_SAMPLE_ID <- sample_name
         if(tier == 'tier3B'){
           tierset$TIER_DESCRIPTION <- 'Tier 3B - Unclassified variants'
@@ -762,24 +761,24 @@ generate_tier_tsv_cpsr <- function(pcg_report, sample_name = "test"){
   tsv_variants$PATHDOC <- stringr::str_replace_all(tsv_variants$PATHDOC,"<br>-", ",")
   tsv_variants <- tsv_variants %>% dplyr::distinct()
 
-  pcg_report[['snv_indel']][['variant_set']][['tsv']] <- tsv_variants
+  cps_report[['snv_indel']][['variant_set']][['tsv']] <- tsv_variants
   for(t in c('TIER 1','TIER 2','TIER 3A','TIER 3B','GWAS')){
     if(t == 'TIER 1'){
-      pcg_report[['snv_indel']][['variant_set']][['tier1']] <- dplyr::filter(tsv_variants, TIER == 'TIER 1')
+      cps_report[['snv_indel']][['variant_set']][['tier1']] <- dplyr::filter(tsv_variants, TIER == 'TIER 1')
     }
     if(t == 'TIER 2'){
-      pcg_report[['snv_indel']][['variant_set']][['tier2']] <- dplyr::filter(tsv_variants, TIER == 'TIER 2')
+      cps_report[['snv_indel']][['variant_set']][['tier2']] <- dplyr::filter(tsv_variants, TIER == 'TIER 2')
     }
     if(t == 'TIER 3A'){
-      pcg_report[['snv_indel']][['variant_set']][['tier3A']] <- dplyr::filter(tsv_variants, TIER == 'TIER 3A')
+      cps_report[['snv_indel']][['variant_set']][['tier3A']] <- dplyr::filter(tsv_variants, TIER == 'TIER 3A')
     }
     if(t == 'TIER 3B'){
-      pcg_report[['snv_indel']][['variant_set']][['tier3B']] <- dplyr::filter(tsv_variants, TIER == 'TIER 3B')
+      cps_report[['snv_indel']][['variant_set']][['tier3B']] <- dplyr::filter(tsv_variants, TIER == 'TIER 3B')
     }
     if(t == 'GWAS'){
-      pcg_report[['snv_indel']][['variant_set']][['gwas']] <- dplyr::filter(tsv_variants, TIER == 'GWAS')
+      cps_report[['snv_indel']][['variant_set']][['gwas']] <- dplyr::filter(tsv_variants, TIER == 'GWAS')
     }
   }
-  return(pcg_report)
+  return(cps_report)
 }