+++ docs/CHANGELOG.md
@@ -1,9 +1,10 @@
 
 ## CHANGELOG
 
-#### dev/unreleased - Nov 6th 2018
+#### dev/unreleased - Nov 10th 2018
 
 ##### Fixed
+ * Bug in assignment of variants to tier1/tier2 [Issue #61](https://github.com/sigven/pcgr/issues/61)
  * Missing config option for *maf_gnomad_asj* in TOML file (also setting operator to `<=`) [Issue #60](https://github.com/sigven/pcgr/issues/60)
  * Bug in new CancerMine oncogene/tumor suppressor annotation [Issue #53](https://github.com/sigven/pcgr/issues/53)
  * vcfanno fix for empty Description (upgrade to vcfanno v0.3.1 [Issue #49](https://github.com/sigven/pcgr/issues/49))